Update to fix Arduino Leonardo/Micro uploads by accounting for variable reset timing on Atmega32u4 based controllers
Originally, the `ButterflyBootloaderProgrammer` uses the 1200bps pulse
trick to for the device into bootloader mode, then waits 1000ms, then
check the com ports again to look for the new programmer port. This does
not work on all computers because some devices do not reset after 1000ms
(they might reset after 1500ms for example), so ArduinoSketchUploader
will fail to upload the new firmware in that case.

This PR fixes that by replacing the 1000ms check with a loop that runs
every 100ms until the new port is found, or timeout is reached (10
seconds), this is much more flexible and allows for devices with a
variety of reset times to work with ArduinoSketchUploader